+++ src/ziozio/dto/ClothWithColor.java
@@ -4,8 +4,8 @@ public class ClothWithColor extends Cloth {
 	private int color_code;
 	private String color_name;
 	private int color_hue_rotate;
-	private double color_saturate;
-	private double color_brightness;
+	private int color_saturate;
+	private int color_brightness;
 	@Override
 	public String toString() {
 		return "ClothWithColor [color_code=" + color_code + ", color_name=" + color_name + ", color_hue_rotate="
@@ -30,22 +30,16 @@ public class ClothWithColor extends Cloth {
 	public void setColor_hue_rotate(int color_hue_rotate) {
 		this.color_hue_rotate = color_hue_rotate;
 	}
-	public double getColor_saturate() {
+	public int getColor_saturate() {
 		return color_saturate;
 	}
-	public void setColor_saturate(double color_saturate) {
+	public void setColor_saturate(int color_saturate) {
 		this.color_saturate = color_saturate;
 	}
-	public double getColor_brightness() {
+	public int getColor_brightness() {
 		return color_brightness;
 	}
-	public void setColor_brightness(double color_brightness) {
+	public void setColor_brightness(int color_brightness) {
 		this.color_brightness = color_brightness;
 	}
-	
-
-
-
-	
-	
 }
